+++ akka/src/test/scala/io/kagera/akka/PetriNetInstanceSpec.scala
@@ -8,9 +8,10 @@ import akka.util.Timeout
 import io.kagera.akka.AkkaTestBase.GetChild
 import io.kagera.akka.actor.PetriNetInstance
 import io.kagera.akka.actor.PetriNetInstanceProtocol._
-import io.kagera.api.colored.ExceptionStrategy.{ BlockTransition, Fatal, RetryWithDelay }
+import io.kagera.execution.ExceptionStrategy.{ BlockTransition, Fatal, RetryWithDelay }
 import io.kagera.api.colored._
 import io.kagera.api.colored.dsl._
+import io.kagera.execution.TransitionExceptionHandler
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.mock.MockitoSugar
 import org.scalatest.time.{ Milliseconds, Span }
