+++ api/src/main/scala/io/kagera/execution/ExceptionState.scala
@@ -1,7 +1,5 @@
 package io.kagera.execution
 
-import io.kagera.api.colored.ExceptionStrategy
-
 case class ExceptionState(
   failureTime: Long,
   failureCount: Int,
